Call setTabPath when creating item group
When trying to use datagen to generate recipes for items that are in the obsidian expansion tab, there is an exception because the tab has a capital letter in the label and does not set its tab path to a string using only lowercase characters (see `ItemGroup#BUILDING_BLOCKS` for an example). This PR fixes the issue by simply calling `setTabPath` and passing a lowercase version of the tab label.